dashboard/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.16.0
Pillow>=9.0.0
